+++ src/enqueue_places.js
@@ -3,9 +3,10 @@ const Apify = require('apify');
 const querystring = require('querystring');
 
 const Puppeteer = require('puppeteer'); // eslint-disable-line
-const { SearchResultOutcome, ScrapingOptions } = require('./typedefs'); // eslint-disable-line no-unused-vars
+const typedefs = require('./typedefs'); // eslint-disable-line no-unused-vars
 const Stats = require('./stats'); // eslint-disable-line no-unused-vars
 const PlacesCache = require('./places_cache'); // eslint-disable-line no-unused-vars
+const MaxCrawledPlacesTracker = require('./max-crawled-places'); // eslint-disable-line no-unused-vars
 
 const { sleep, log } = Apify.utils;
 const { PLACE_TITLE_SEL, NEXT_BUTTON_SELECTOR, NO_RESULT_XPATH } = require('./consts');
@@ -15,24 +16,25 @@ const { checkInPolygon } = require('./polygon');
 
 const SEARCH_WAIT_TIME_MS = 30000;
 const CHECK_LOAD_OUTCOMES_EVERY_MS = 500;
-const PLACES_PER_PAGE = 20;
 
 /**
  * This handler waiting for response from xhr and enqueue places from the search response boddy.
  * @param {{
  *   page: Puppeteer.Page,
  *   requestQueue: Apify.RequestQueue,
+ *   request: Apify.Request,
  *   searchString: string,
- *   maxPlacesPerCrawl: number | undefined,
  *   exportPlaceUrls: boolean,
  *   geo: object,
  *   placesCache: PlacesCache,
  *   stats: Stats,
+ *   maxCrawledPlacesTracker: MaxCrawledPlacesTracker,
  * }} options
  * @return {(response: Puppeteer.Response) => Promise<void>}
  */
 const enqueuePlacesFromResponse = (options) => {
-    const { page, requestQueue, searchString, maxPlacesPerCrawl, exportPlaceUrls, geo, placesCache, stats } = options;
+    const { page, requestQueue, searchString, request, exportPlaceUrls, geo,
+        placesCache, stats, maxCrawledPlacesTracker } = options;
     return async (response) => {
         const url = response.url();
         if (url.match(/google\.[a-z.]+\/search/)) {
@@ -51,37 +53,48 @@ const enqueuePlacesFromResponse = (options) => {
             for (const placePaginationData of placesPaginationData) {
                 index++;
                 const rank = ((pageNumber - 1) * 20) + (index + 1);
-                if (!maxPlacesPerCrawl || rank <= maxPlacesPerCrawl) {
-                    if (exportPlaceUrls) {
-                        await Apify.pushData({
-                            url: `https://www.google.com/maps/search/?api=1&query=${searchString}&query_place_id=${placePaginationData.placeId}`,
-                        });
-                    } else {
-                        // TODO: Refactor this once we get rid of the caching
-                        const coordinates = placePaginationData.coords || placesCache.getLocation(placePaginationData.placeId);
-                        const placeUrl = `https://www.google.com/maps/search/?api=1&query=${searchString}&query_place_id=${placePaginationData.placeId}`;
-                        placesCache.addLocation(placePaginationData.placeId, coordinates, searchString);
-                        if (!geo || !coordinates || checkInPolygon(geo, coordinates)) {
-                            await requestQueue.addRequest({
-                                    url: placeUrl,
-                                    uniqueKey: placePaginationData.placeId,
-                                    userData: {
-                                        label: 'detail',
-                                        searchString,
-                                        rank,
-                                        searchPageUrl,
-                                        coords: placePaginationData.coords,
-                                        addressParsed: placePaginationData.addressParsed,
-                                        isAdvertisement: placePaginationData.isAdvertisement,
-                                    },
+                if (exportPlaceUrls) {
+                    await Apify.pushData({
+                        url: `https://www.google.com/maps/search/?api=1&query=${searchString}&query_place_id=${placePaginationData.placeId}`,
+                    });
+                } else {
+                    // TODO: Refactor this once we get rid of the caching
+                    const coordinates = placePaginationData.coords || placesCache.getLocation(placePaginationData.placeId);
+                    const placeUrl = `https://www.google.com/maps/search/?api=1&query=${searchString}&query_place_id=${placePaginationData.placeId}`;
+                    placesCache.addLocation(placePaginationData.placeId, coordinates, searchString);
+                    if (!geo || !coordinates || checkInPolygon(geo, coordinates)) {
+                        const searchKey = searchString || request.url;
+                        if (!maxCrawledPlacesTracker.setEnqueued(searchKey)) {
+                            log.warning(`[SEARCH]: Finishing search because we enqueued more than maxCrawledPlaces `
+                                + `currently: ${maxCrawledPlacesTracker.enqueuedPerSearch[searchKey]}(for this search)/${maxCrawledPlacesTracker.enqueuedTotal}(total) `
+                                + `--- ${searchString} - ${request.url}`);
+                            break;
+                        }
+                        const { wasAlreadyPresent } = await requestQueue.addRequest({
+                                url: placeUrl,
+                                uniqueKey: placePaginationData.placeId,
+                                userData: {
+                                    label: 'detail',
+                                    searchString,
+                                    rank,
+                                    searchPageUrl,
+                                    coords: placePaginationData.coords,
+                                    addressParsed: placePaginationData.addressParsed,
+                                    isAdvertisement: placePaginationData.isAdvertisement,
                                 },
-                                { forefront: true });
+                            },
+                            { forefront: true });
+                        if (!wasAlreadyPresent) {
                             enqueued++;
                         } else {
-                            stats.outOfPolygonCached();
-                            stats.outOfPolygon();
-                            stats.addOutOfPolygonPlace({ url: placeUrl, searchPageUrl, coordinates });
+                            log.warning(`Google presented already enqueued place, skipping... --- ${placeUrl}`)
+                            maxCrawledPlacesTracker.enqueuedTotal--;
+                            maxCrawledPlacesTracker.enqueuedPerSearch[searchKey]--;
                         }
+                    } else {
+                        stats.outOfPolygonCached();
+                        stats.outOfPolygon();
+                        stats.addOutOfPolygonPlace({ url: placeUrl, searchPageUrl, coordinates });
                     }
                 }
             }
@@ -95,7 +108,7 @@ const enqueuePlacesFromResponse = (options) => {
 /**
  * Periodically checks if one of the possible search outcomes have happened
  * @param {Puppeteer.Page} page
- * @returns {Promise<SearchResultOutcome>} // Typing this would require to list all props all time
+ * @returns {Promise<typedefs.SearchResultOutcome>} // Typing this would require to list all props all time
  */
 const waitForSearchResults = async (page) => {
     const start = Date.now();
@@ -142,8 +155,8 @@ const waitForSearchResults = async (page) => {
  *  searchString: string,
  *  requestQueue: Apify.RequestQueue,
  *  request: Apify.Request,
- *  stats: Stats,
- *  scrapingOptions: ScrapingOptions,
+ *  helperClasses: typedefs.HelperClasses,
+ *  scrapingOptions: typedefs.ScrapingOptions,
  * }} options
  */
 module.exports.enqueueAllPlaceDetails = async ({
@@ -151,20 +164,22 @@ module.exports.enqueueAllPlaceDetails = async ({
                                           searchString,
                                           requestQueue,
                                           request,
-                                          stats,
                                           scrapingOptions,
+                                          helperClasses,
                                       }) => {
-    const { geo, maxAutomaticZoomOut, placesCache, exportPlaceUrls, maxCrawledPlaces } = scrapingOptions;
+    const { geo, maxAutomaticZoomOut, exportPlaceUrls } = scrapingOptions;
+    const { stats, placesCache, maxCrawledPlacesTracker } = helperClasses;
 
     page.on('response', enqueuePlacesFromResponse({
         page,
         requestQueue,
         searchString,
-        maxPlacesPerCrawl: maxCrawledPlaces,
+        request,
         exportPlaceUrls,
         geo,
         placesCache,
         stats,
+        maxCrawledPlacesTracker,
     }));
     // there is no searchString when startUrls are used
     if (searchString) {
@@ -178,8 +193,6 @@ module.exports.enqueueAllPlaceDetails = async ({
     await waitForGoogleMapLoader(page);
 
     const startZoom = /** @type {number} */ (parseZoomFromUrl(page.url()));
-
-    let pagesScraped = 0;
 
     for (;;) {
         const {
@@ -210,8 +223,8 @@ module.exports.enqueueAllPlaceDetails = async ({
             return;
         }
 
-        if (maxCrawledPlaces && pagesScraped * PLACES_PER_PAGE > maxCrawledPlaces) {
-            log.warning(`[SEARCH]: Finishing search because we scraped more than maxCrawledPlaces --- ${searchString} - ${request.url}`);
+        if (!maxCrawledPlacesTracker.canEnqueueMore(searchString || request.url)) {
+            // no need to log here because it is logged already in enqueuePlacesFromResponse
             return;
         }
 
@@ -238,6 +251,5 @@ module.exports.enqueueAllPlaceDetails = async ({
             await page.evaluate((sel) => $(sel).click(), NEXT_BUTTON_SELECTOR);
             await waitForGoogleMapLoader(page);
         }
-        pagesScraped++;
     }
 };